+++ android/src/main/kotlin/com/smileidentity/flutter/enhanced/SmileIDSmartSelfieAuthenticationEnhanced.kt
@@ -5,6 +5,7 @@ import androidx.compose.runtime.Composable
 import com.smileidentity.SmileID
 import com.smileidentity.compose.SmartSelfieAuthenticationEnhanced
 import com.smileidentity.flutter.SmileComposablePlatformView
+import com.smileidentity.flutter.SmileSelfieComposablePlatformView
 import com.smileidentity.networking.FileAdapter
 import com.smileidentity.results.SmartSelfieResult
 import com.smileidentity.results.SmileIDResult
@@ -21,7 +22,7 @@ internal class SmileIDSmartSelfieAuthenticationEnhanced private constructor(
     viewId: Int,
     messenger: BinaryMessenger,
     args: Map<String, Any?>,
-) : SmileComposablePlatformView(context, VIEW_TYPE_ID, viewId, messenger, args) {
+) : SmileSelfieComposablePlatformView(context, VIEW_TYPE_ID, viewId, messenger, args) {
     companion object {
         const val VIEW_TYPE_ID = "SmileIDSmartSelfieAuthenticationEnhanced"
     }
@@ -34,38 +35,10 @@ internal class SmileIDSmartSelfieAuthenticationEnhanced private constructor(
             allowNewEnroll = args["allowNewEnroll"] as? Boolean ?: false,
             showAttribution = args["showAttribution"] as? Boolean ?: true,
             showInstructions = args["showInstructions"] as? Boolean ?: true,
+            skipApiSubmission = args["skipApiSubmission"] as? Boolean ?: false,
             extraPartnerParams = extraPartnerParams.toImmutableMap(),
-        ) {
-            val moshi =
-                Moshi
-                    .Builder()
-                    .add(FileAdapter)
-                    .build()
-            when (it) {
-                is SmileIDResult.Success -> {
-                    val result =
-                        SmartSelfieResult(
-                            selfieFile = it.data.selfieFile,
-                            livenessFiles = it.data.livenessFiles,
-                            apiResponse = it.data.apiResponse,
-                        )
-                    val json =
-                        try {
-                            moshi
-                                .adapter(SmartSelfieResult::class.java)
-                                .toJson(result)
-                        } catch (e: Exception) {
-                            onError(e)
-                            return@SmartSelfieAuthenticationEnhanced
-                        }
-                    json?.let { response ->
-                        onSuccessJson(response)
-                    }
-                }
-
-                is SmileIDResult.Error -> onError(it.throwable)
-            }
-        }
+            onResult = { res -> handleResult(res) },
+        )
     }
 
     class Factory(