+++ ios/Classes/SmileIDSmartSelfieEnrollmentEnhanced.swift
@@ -29,6 +29,7 @@ class SmileIDSmartSelfieEnrollmentEnhanced: NSObject, FlutterPlatformView, Smart
             allowNewEnroll: args["allowNewEnroll"] as? Bool ?? false,
             showAttribution: args["showAttribution"] as? Bool ?? true,
             showInstructions: args["showInstructions"] as? Bool ?? true,
+            skipApiSubmission: args["skipApiSubmission"] as? Bool ?? false,
             extraPartnerParams: args["extraPartnerParams"] as? [String: String] ?? [:],
             delegate: self
         )
@@ -109,6 +110,7 @@ struct EnhancedSelfieEnrollmentRootView: View {
     let allowNewEnroll: Bool
     let showAttribution: Bool
     let showInstructions: Bool
+    let skipApiSubmission: Bool
     let extraPartnerParams: [String: String]
     let delegate: SmartSelfieResultDelegate
 
@@ -119,6 +121,7 @@ struct EnhancedSelfieEnrollmentRootView: View {
                 allowNewEnroll: allowNewEnroll,
                 showAttribution: showAttribution,
                 showInstructions: showInstructions,
+                skipApiSubmission: skipApiSubmission,
                 extraPartnerParams: extraPartnerParams,
                 delegate: delegate
             )